Update patch set 1
Patch Set 1:

> you could (and should afaict) clean up |aes_init_key| a bit.

Are you suggesting that `aes_init_key()` can be simplified by assuming that BSAES and VPAES are mutually-exclusive? I think it might make sense to make that simplification, but if we were to do so then we'd probably be better off adding an explicit check somewhere that BSAES and VPAES actually are mutually-exclusive.

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>