Update patch set 9
Patch Set 9:

OK, I remember why I didn't want to do things the way Patchset 9 does it: It is possible that the compiler will allocate the 4,480-byte buffer on the stack before calling into the RSAZ code. Then the RSAZ code would allocate its 5.5KB buffer on the stack, resulting in 10KB of stack usage.

In Patchset 8, that won't happen, but the cost of that not happening is an extra ~1KB (5.5KB - 4,480 bytes) of stack space wasted when the non-RSAZ code path is taken.

So I think Patchset 8 is better. LMK what you think.

Patch-set: 9