Update patch set 2
Patch Set 2:

This is mostly to start a discussion. IIUC the code in question is actually the output from Fiat Crypto and we should change Fiat Crypto and/or move this change into the definitions of |cmovznz64| and |cmovznz32| if they are out of the scope of the verification.

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>