Update patch set 2
Patch Set 2:

> This is mostly to start a discussion. IIUC the code in question is actually the output from Fiat Crypto

It is, yeah.

> we should change Fiat Crypto and/or move this change into the definitions of |cmovznz64| and |cmovznz32| if they are out of the scope of the verification.

cmovznz* are written by hand and unverified, yeah. Of course, the largely mechanical translation from fiat-crypto's actual IR to C is also unverified, but still I think that warrants avoiding modifications to those callers, especially while the intrinsics are written separately from fiat-crypto. (Ideally they would eventually output something fully standalone, including any intrinsics since that's ultimately tied to the C translation, so that this stuff wouldn't have to be integrated quite so manually.)

We could see if fiat-crypto would be interested in applying an extra transformation for cmovznz(?, 0, 0xffff...), but P-256 has more general ones, so this doesn't remove the need for this intrinsic in general. So if it works and the compiler handles the optimization, it seems fine enough? 

It's true that this code, like everything else, is at the mercy of the compiler and what happens to produce constant-time code. Ultimately the answer needs to be getting some actual constant-time support in there, but I think compiler folks are all a bit tied up with speculative execution excitement right now. :-)

A thought: the cmov stuff does predate fiat-crypto doing a better job of dealing with 1-bit values. Previously they rounded up to 8-bit at the wrong time, which is I think why the cmovs are a bit silly. (Hence the uint8_t /* bool */ stuff which means it is either zero or one.) I suspect the actual type of cmov should be uint8_t /* bool */ which means you can get rid of the !!. I think that would do basically what you want here?

(Yes, this is all very silly and we should get the intrinsics emitted as part of fiat-crypto in one go.)

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>