Update patch set 4
Patch Set 4:

> We could see if fiat-crypto would be interested in applying an extra transformation for cmovznz(?, 0, 0xffff...), but P-256 has more general ones, so this doesn't remove the need for this intrinsic in general. So if it works and the compiler handles the optimization, it seems fine enough? 

I think it's asking too much to expect the compiler to optimize `!!t` exactly the right way and to do it in a branch-free manner.

> A thought: the cmov stuff does predate fiat-crypto doing a better job of dealing with 1-bit values. Previously they rounded up to 8-bit at the wrong time, which is I think why the cmovs are a bit silly. (Hence the uint8_t /* bool */ stuff which means it is either zero or one.) I suspect the actual type of cmov should be uint8_t /* bool */ which means you can get rid of the !!. I think that would do basically what you want here?

I made a change similar to that. OTOH, if one wants to avoid logical operators then the compiler might generate less optimal code since it might not see that it should use the carry flag. Thus, I'm not sure this new version is better in terms of performance.

Patch-set: 4
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>