Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> David: can you review? Since I rewrote this, it's effectively unreviewed at this point.
> 
> Nir, Shay: is the current state of this change good? (I.e. with my changes to use memcpy explicitly to avoid any undefined behaviour?)

Hi Adam,
 
We benchmarked your modified code (the one that uses OPENSSL_memcpy).
 
Introducing memcopy instead of jmp is not so good. It actually neutralizes the performance gains, and makes the result more or less the same as before the patch. Probably because the changed code has many "call" instructions (to OPENSSL_memcpy) that replace the previous "jmp" instructions. 
 
How about the following idea. The original commit was tested on X86_64 machines (several; all show the improvement). So, why not narrow down the optimization to these architectures (until there is a comprehensive solution that fits all).
 
That is, can you revive the original optimization (latest commit), but: to mitigate the concern that other platforms (ARM?) will not handle the type casting, wrap the changed code with "#ifdef OPENSSL_X86_64". Then, we can enjoy the targeted performance without worrying about side effects on other platforms. 
 
Thanks,
Nir and Shay

Patch-set: 3