Update patch set 3
Patch Set 3:

> We therefore recommend to use the optimization only for AVX512 (i.e., under some #ifdef AVX512)

If I'm understanding you correctly, you suggest not bothering with the C code, but using the AVX-512 version?

However, the AVX-512 code would have to be in an assembly file, and be guarded by a CPUID-based check. The compiler cannot inline across objects, so we would end up having a function calls which might well push that improvement into the noise too :(

(Maybe it can be done as intrinsics, but AVX-512 intrinsics will only be supported by the newest compilers, so we'll have to figure out the correct #define magic not to break builds.)

So it's possible, but it seems marginal and not clearly worthwhile.

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Commit-Queue=0