Move "prediction resistance" later in RAND_bytes_with_additional_data.
Reduce the time window in which a VM clone would result in a duplicate
state. This adds an additional lock/unlock in the uncommon case where a
reseeding is necessary; this shouldn't have any material impact on
performance in FIPS mode and won't have any negative performance effect
in the good mode.

Change-Id: I38fac61242fe5924080ab4d4dc8216b156ddb193