Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Raise 2*R to the r_bits power to get 2^r_bits*R = R*R.
> 
> You can also use shifts to calculate 4*R or 8*R or 16*R using shifts, cutting the exponent in half/4/8 accordingly, if shifting is faster than squaring. I don't know what the shift:square trade-off is though.

Oh, good point! I'd toyed with 2R since that avoids worrying about the value of r_bits, but r_bits is actually a multiple of 32 or 64, so we can get all the way up to 2^32*R or 2^64*R (platform-specific) if we wanted.

Though I believe N additional shifts pays for log_2(N+1) square/mults, so it'd probably stop being useful long before we get to 2^32.

Patch-set: 5