Expose certificate compression information to the application
This adds a new API to expose certificate compression information to
applications (compression algorithm used, uncompressed length and
compressed length). This is useful for gathering metrics on certificate
compression.

It would probably be possible to do the same by exposing this information
via e.g. ex_data from the compression callbacks themselves, but this
seems easier and it can be re-used across applications.

Change-Id: Id8f23b5f3692bf9104e807f255d558cb371e6c21