Serialize SSL configuration in handback and check it on application.
A split SSL handshake involves 3 |SSL| objects: the initial |SSL|
receives the |ClientHello| message, the handoff |SSL| performs some or
all of the handshake, and the handback |SSL| finishes the handshake,
if it is not already finished, and passes traffic.

The configuration of the initial |SSL| is irrelevant, since it is not
responsible for sending any traffic to the client.

But we would like to guarantee that the handoff |SSL| does not make
any promises that the handback |SSL| cannot keep.  This commit adds
the following to the handback message, from the handoff |SSL|'s
configuration:

* (min version, max version)
* List of configured ciphers
* List of supported curves

SSL_apply_handback() is modified to reject any handback message that
indicates that the handoff |SSL| was explicitly or implicitly
configured to support something not supported by the handback |SSL|.

Change-Id: Idf6dbeadb750c076ab0509c09b9d3f22eb162b9c