Update patch set 2
Patch Set 2:

(9 comments)

I started reviewing this, but I think we were actually talking about different things. :-)

This CL makes it so that, if the handshaker *supports* anything that the edge doesn't support, it breaks. But we don't want that. We want it to break if the handshaker *selects* anything if the edge doesn't support, but I would hope that happens anyway by way of it rejecting code points it doesn't recognize.

But once we get to that point, something's already gone wrong because we've broken the connection. Rather, we want it to be the case that the handshaker never *selects* anything the edge can't handle. A priori, that means we must carefully handle deployment so that we never turn anything on in the handshaker until support code is deployed everywhere to edge. Conversely, we never remove support for something at the edge until it is turned off in the handhaker. However, that seems easy to mess up.

To alleviate the common case, I was proposing that the hand*off* message include a list of code points the edge is prepared to deal with. The handshaker then, in the actual negotiation logic, accounts for this and never selects a code point that doesn't match. It does make the protocol more chatty, but hopefully not too much? More complex stuff we'll probably still need to take care, or perhaps exercise the version field, but this should cover the common case.

My other comment is that this is based on the edge configuration rather than the edge's code. I was actually envisioning using the code, but maybe config is better? The question is what should the edge caller be obligated to configure. Note the certificate and key material, presumably. What about other bits of config? Should it configure ciphers, or should it leave that to the handshaker [up to what's supported in code]? I've been assuming the latter, since having two places to coordinate configuration seems worse than one, but maybe there is interest in the former.

Thoughts?

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>