Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

Thanks! It looks like this is a CMake bug. This line here is missing quotes. If I patch that in, (see &quot; instances elsewhere in the file), the build works.
https://github.com/Kitware/CMake/blob/master/Templates/MSBuild/nasm.xml#L51

I'll send CMake a patch to clear this going forward, but, in the meantime, we can carry this as a workaround.

Patch-set: 3
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+1