Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (9 comments)
> 
> I started reviewing this, but I think we were actually talking about different things. :-)

I make some arguments below for my approach.  I'm not wedded to them, so do let me know if you find them unconvincing.

> This CL makes it so that, if the handshaker *supports* anything that the edge doesn't support, it breaks. But we don't want that. We want it to break if the handshaker *selects* anything if the edge doesn't support, but I would hope that happens anyway by way of it rejecting code points it doesn't recognize.

If we enable (say) a new experimental ciphersuite on the handshaker, without enabling it on the edge, that error could go unnoticed until some client selects the new ciphersuite.  So, I would argue, things should break when the mere potential for an erroneous selection exists, because that closely couples the breakage signal to the production change that ought to be reverted.

> But once we get to that point, something's already gone wrong because we've broken the connection. Rather, we want it to be the case that the handshaker never *selects* anything the edge can't handle. A priori, that means we must carefully handle deployment so that we never turn anything on in the handshaker until support code is deployed everywhere to edge. Conversely, we never remove support for something at the edge until it is turned off in the handhaker. However, that seems easy to mess up.
> 
> To alleviate the common case, I was proposing that the hand*off* message include a list of code points the edge is prepared to deal with. 

The edge comprises many machines, running different code and having different configs.  So that scheme could lead to a situation where the handoff message from edge machine 1 commits to something that edge machine 2, the receiver of the handback message, is unable to handle.

And if that happens, it will likely happen with low probability, making it hard to diagnose.

> The handshaker then, in the actual negotiation logic, accounts for this and never selects a code point that doesn't match. It does make the protocol more chatty, but hopefully not too much? More complex stuff we'll probably still need to take care, or perhaps exercise the version field, but this should cover the common case.

> My other comment is that this is based on the edge configuration rather than the edge's code. I was actually envisioning using the code, but maybe config is better? 

The benefit of using config rather than code is that the requirement to serialize edge and handshaker deployments kicks in only when the operator enables a new feature explicitly.  If you don't enable any new features, you can deploy however you want.

(Although this argument is moot if we use your idea of having the first edge machine's config limit the options that may be used downstream.)

> The question is what should the edge caller be obligated to configure. Note the certificate and key material, presumably. What about other bits of config? Should it configure ciphers, or should it leave that to the handshaker [up to what's supported in code]? I've been assuming the latter, since having two places to coordinate configuration seems worse than one, but maybe there is interest in the former.

It is true that my scheme would require the edge to explicitly enable any feature that is not enabled by default, before that feature can be enabled on the handshaker.

Your scheme has fewer moving parts but mine is more explicit.

For example, in your scheme, a new cipher could be first enabled on the handshaker, but it would not be used until support for the new cipher arrived in the edge binary.  One could avoid this with good rollout discipline (i.e. always deploying new code to the edge first), but the possibility feels weird.

Whereas in my scheme, provided that the new cipher is not enabled by default, enabling the new cipher in the handshaker would cause an error until the new cipher code arrives in the edge binary, *and* the edge is explicitly configured to support the new cipher.

The explicit-configuration requirement for the edge could be avoided: instead, the handshaker could pass its full list of configured ciphers to the edge, with the understanding that the edge would reject the handback if the cipher list contained anything it didn't understand, and would accept any cipher that it did understand, regardless of how its cipher list was configured by the caller.

Patch-set: 2
Reviewer: Gerrit User 5645 <5645@66b711a4-aa9b-3dac-9062-916540830212>
Label: Commit-Queue=0