Add SHA-256×16 – a j-lanes SHA-256.
This change only has optimised code for x86-64 on Linux. We could
check-in versions for macOS and Windows too, but it's unclear whether
the compiler-generated code is actually going to cut it, performance
wise. We might need to hand-write it anyway, at which point we can just
use perlasm.

For now, however, the instrinsics code is certainly faster than the
generic code and Linux is the primary OS of interest for this.

Change-Id: I110d3ad090f504c650a9907a041c6e7928ed4aab