Update patch set 2
Patch Set 2:

I think the motivation here is that you want to run the full set of runner tests in cross-version configurations, right? And so you need to be able to MoveTestState across processes?

I may have suggested doing exactly this in the past but, seeing this code, I do worry that making the test state a semi-stable interface is going to be too much burden and wonder whether we should have different, less intrusive tests for cross-version testing.

It's good to do the runner tests in split-handshake mode within a version but we often add features to the shim to enable new tests. Those tests will inevitable be broken in a cross-version context because the older versions won't support the feature that we're testing. But, if the cross-version tests aren't to be red all the time, how is that to work?

We could start tagging tests with the minimum version that the handshaker needs to support in order to function, but that adds a concept of versions in BoringSSL that we've never had before.

Alternatively, maybe we need a different set of tests, that still use the shim for cross-version testing, but which check behaviour in a black-box manner, and which lag the runner tests in terms of testing features?

(I suspect you've thought more about how this will work and I should stop prattling on.)

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>