Update patch set 1
Patch Set 1:

Oh, I can make one request to start with:

This CL is taking a master-with-bazel commit and trying to submit it to master. That's not quite right. The master-with-bazel branch is automatically generated from master.

Instead, please clone the master branch and make your change to util/BUILD.toplevel? That file becomes the top-level BUILD file in the master branch. If you copy the Change-Id line from your commit message, it should upload to the same CL.

Patch-set: 1