Update patch set 3
Patch Set 3:

I've noodled around with this a bit; PTAL.  To summarize the changes:

We can't avoid serializing sessions if we want to test all the important cases, since sessions are written by the shim and consumed by the handshaker.

So what else is still serialized?  Serializing the ticket keys could be avoided by setting them to some constant value when otherwise unspecified.  I've continued to serialize the handshake transcript, because it's high-value, but we could if we chose treat it the same as |early_callback_called|, i.e. don't look at it in the config->handoff case.

Does this represent a sufficiently blackbox-y happy medium?

Patch-set: 3
Reviewer: Gerrit User 5645 <5645@66b711a4-aa9b-3dac-9062-916540830212>