Update patch set 3
Patch Set 3:

> Patch Set 2: Commit-Queue+1
> 
> DO NOT SUBMIT.
> 
> (At least, not just before M70 branches.)
> 
> Just an idea.

Going to put that in the commit message otherwise I'm probably going to forget. :-)

We still have a bit of time for M70, but M70's carrying the final TLS 1.3 so maybe not in the same release. There may also be something to be said for hooking this one up to Finch and stuff, or at least some probing to make sure it's safe.

Patch-set: 3