Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Ooops, hit enter too fast.
> > 
> > What's the plan for keeping that list up-to-date? If we're doing something source-based, we'll definitely need one, since we can't reasonably manually update that. It would also be preferable if generating the list were automated so we don't have a lot of diff noise every time we add a function.
> 
> I was actually thinking that we'd require manual updating, but provide an automated test to compile two versions of the build artifacts with two different prefixes and then compare them to make sure that there's no overlap between the symbols. That way, you'd always catch people who forgot, and once you know to update the list, doing so takes 30 seconds.

That still sounds like a huge amount of friction on development. Post-facto testing allows moving from "automated" to "mostly automated", but I think the "mostly" is still important.

> > The C++ bits need an inline namespace, of course, but we can do that separately. That one's really easy.
> > 
> > > Note that, right now, all of the symbols - including internal ones - are handled in a single, public header file. This could hypothetically conflict with the internals of consumers' implementations. Let me know if you want me to look into moving the internal-only stuff into internal header files.
> > 
> > I think that's fine. I'm more worried about symbols like HMAC.
> 
> I don't follow - don't you /want/ to rename public symbols in consumer code so that it matches what ends up in BoringSSL's object files?

Yes. But because of symbols like HMAC, this macro strategy is basically unusable for anything but uses like yours, where you're not actually using the macros in consumer code. This is why we abandoned this approach originally. C macros are not hygienic. If we have an HMAC macro, then folks with enums or namespaced C++ symbols or whatever break. (That is in fact exactly the symbol that frustrated this strategy last time.)

In contrast, internal symbols are less of a concern here because we can freely rename them to suitably start with "crypto_" or "evp_" or whatever.

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>